requirements.txt
annotated-types==0.7.0
ansible-runner==2.4.0
anyio==4.4.0
certifi==2024.6.2
docutils==0.21.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.7
lockfile==0.12.2
onepassword==0.1.0b1
onepasswordconnectsdk==1.5.1
packaging==24.1
pexpect==4.9.0
ptyprocess==0.7.0
pydantic==2.7.4
pydantic_core==2.18.4
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rfc3986==1.5.0
six==1.16.0
slack_sdk==3.29.0
sniffio==1.3.1
typing_extensions==4.12.2
